requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-crispy-forms==1.14.0
django-environ==0.4.5
django-recaptcha==2.0.6
google-auth==2.6.0
gunicorn==20.1.0
idna==3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
